MIFT-108 add ability to view friends calendar, MIFT-136 MIFT-137 bug fixes

+++ src/user-event/userEvent.handler.ts
@@ -10,14 +10,36 @@ import { UserEvent } from './userEvent.model';
 import { convertToUTC } from '../helpers/time';
 import { AuthenticatedRequest } from '../domain/authenticatedRequest';
 import { GetUserLectureEventsRequestDto } from '../lecture-event/dto/request/get-user-lecture-events.dto';
+import { User } from '../user/user.model';
 
 export const fetchUserEvents = catchAsync(
   async (req: AuthenticatedRequest, res, next) => {
-    const { startDateTime, endDateTime }: GetUserLectureEventsRequestDto =
-      req.body;
+    const {
+      startDateTime,
+      endDateTime,
+      userId,
+    }: GetUserLectureEventsRequestDto = req.body;
+
+    const foundUser = await User.findById(userId);
+
+    if (!foundUser) {
+      return next(new AppError("The user doesn't exist", 400));
+    }
+
+    if (
+      foundUser.id !== req.user.id &&
+      !req.user.friends.includes(foundUser.id)
+    ) {
+      return next(
+        new AppError(
+          "You can't view the calendars of users who aren't your friends.",
+          400,
+        ),
+      );
+    }
 
     const userEvents = await UserEvent.find({
-      user: req.user._id,
+      user: foundUser._id,
       $and: [
         {
           startDateTime: {
@@ -55,53 +77,24 @@ export const createUserEvent = catchAsync(
       return next(new AppError('Start time must be before end time', 400));
     }
 
-    if (repeatable === true) {
-      let TemporaryDate = new Date(startDateTime);
-      let repeatableUntilDate = new Date(repeatableUntil);
-      let Counter = 0;
-      while (TemporaryDate <= repeatableUntilDate) {
-        let StartDateAsADateObject = new Date(startDateTime);
-        let EndDateAsADateObject = new Date(endDateTime);
-        TemporaryDate.setDate(TemporaryDate.getDate() + 7);
-        StartDateAsADateObject.setDate(
-          StartDateAsADateObject.getDate() + 7 * Counter,
-        );
-        const UpdaterStartDateString = StartDateAsADateObject.toISOString();
-        EndDateAsADateObject.setDate(
-          EndDateAsADateObject.getDate() + 7 * Counter,
-        );
-        const UpdatedEndDateString = EndDateAsADateObject.toISOString();
-        Counter++;
-        await UserEvent.create({
-          startDateTime: convertToUTC(UpdaterStartDateString, req.timezone),
-          endDateTime: convertToUTC(UpdatedEndDateString, req.timezone),
-          title,
-          note,
-          location,
-          repeatable,
-          repeatableUntil,
-          user: req.user,
-        });
-      }
+    const createdUserEvent = await UserEvent.create({
+      startDateTime: convertToUTC(startDateTime, req.timezone),
+      endDateTime: convertToUTC(endDateTime, req.timezone),
+      title,
+      note,
+      location,
+      repeatable,
+      repeatableUntil,
+      user: req.user,
+    });
 
-      res.json({
-        status: ResponseStatus.SUCESS,
-      });
-    } else {
-      const createdUserEvent = await UserEvent.create({
-        startDateTime: convertToUTC(startDateTime, req.timezone),
-        endDateTime: convertToUTC(endDateTime, req.timezone),
-        title,
-        note,
-        location,
-        user: req.user,
-      });
+    req.user.events.push(createdUserEvent.id);
+    await req.user.save();
 
-      res.json({
-        status: ResponseStatus.SUCESS,
-        data: new CreateUserEventResponseDto(createdUserEvent),
-      });
-    }
+    res.json({
+      status: ResponseStatus.SUCESS,
+      data: new CreateUserEventResponseDto(createdUserEvent),
+    });
   },
 );
 
@@ -122,6 +115,12 @@ export const updateUserEvent = catchAsync(
       new Date(updateUserEventDto.endDateTime)
     ) {
       return next(new AppError('Start time must be before end time', 400));
+    }
+
+    if (!req.user.events.includes(userEvent.id)) {
+      console.log(req.user.events);
+      console.log(userEvent.id);
+      return next(new AppError("You can't edit other peoples' events", 400));
     }
 
     //@ts-ignore
@@ -162,6 +161,10 @@ export const deleteUserEvent = catchAsync(async (req, res, next) => {
     return next(
       new AppError('User event with provided ID does not exist', 400),
     );
+  }
+
+  if (!req.user.events.includes(userEvent.id)) {
+    return next(new AppError("You can't delete other peoples' events", 400));
   }
 
   await UserEvent.findOneAndDelete({ _id: userEventId });
